# cbioportal-frontend/packages/cbioportal-ts-api-client/src/generated/CBioPortalAPI-docs.json
{
  "swagger": "2.0",
  "info": {
    "contact": {
      "email": "cbioportal@googlegroups.com",
      "name": "cbioportal",
      "url": "https://www.cbioportal.org"
    },
    "description": "A web service for supplying JSON formatted data to cBioPortal clients. Please note that this API is currently in beta and subject to change.",
    "license": {
      "name": "License",
      "url": "https://github.com/cBioPortal/cbioportal/blob/master/LICENSE"
    },
    "title": "cBioPortal web Public API [Alpha]",
    "version": "1.0 (beta). Backwards compatibility will be maintained (after 1.0 release)"
  },
  "paths": {
...
    "/api/samples": {
      "get": {
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "Search keyword that applies to the study ID",
            "in": "query",
            "name": "keyword",
            "required": false,
            "type": "string"
          },
          {
            "default": "SUMMARY",
            "description": "Level of detail of the response",
            "enum": [
              "ID",
              "SUMMARY",
              "DETAILED",
              "META"
            ],
            "in": "query",
            "name": "projection",
            "required": false,
            "type": "string"
          },
          {
            "default": 10000000,
            "description": "Page size of the result list",
            "format": "int32",
            "in": "query",
            "maximum": 10000000,
            "minimum": 1,
            "name": "pageSize",
            "required": false,
            "type": "integer"
          },
          {
            "default": 0,
            "description": "Page number of the result list",
            "format": "int32",
            "in": "query",
            "minimum": 0,
            "name": "pageNumber",
            "required": false,
            "type": "integer"
          },
          {
            "description": "Name of the property that the result list is sorted by",
            "enum": [
              "sampleId",
              "sampleType"
            ],
            "in": "query",
            "name": "sortBy",
            "required": false,
            "type": "string"
          },
          {
            "default": "ASC",
            "description": "Direction of the sort",
            "enum": [
              "ASC",
              "DESC"
            ],
            "in": "query",
            "name": "direction",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "items": {
                "$ref": "#/definitions/Sample"
              },
              "type": "array"
            }
          }
        },
        "tags": [
          "Samples"
        ],
        "description": "Get all samples matching keyword",
        "operationId": "getSamplesByKeywordUsingGET"
      }
    },
...
    "/api/studies/{studyId}/patients/{patientId}/samples": {
      "get": {
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "Study ID e.g. acc_tcga",
            "in": "path",
            "name": "studyId",
            "required": true,
            "type": "string"
          },
          {
            "description": "Patient ID e.g. TCGA-OR-A5J2",
            "in": "path",
            "name": "patientId",
            "required": true,
            "type": "string"
          },
          {
            "default": "SUMMARY",
            "description": "Level of detail of the response",
            "enum": [
              "ID",
              "SUMMARY",
              "DETAILED",
              "META"
            ],
            "in": "query",
            "name": "projection",
            "required": false,
            "type": "string"
          },
          {
            "default": 10000000,
            "description": "Page size of the result list",
            "format": "int32",
            "in": "query",
            "maximum": 10000000,
            "minimum": 1,
            "name": "pageSize",
            "required": false,
            "type": "integer"
          },
          {
            "default": 0,
            "description": "Page number of the result list",
            "format": "int32",
            "in": "query",
            "minimum": 0,
            "name": "pageNumber",
            "required": false,
            "type": "integer"
          },
          {
            "description": "Name of the property that the result list is sorted by",
            "enum": [
              "sampleId",
              "sampleType"
            ],
            "in": "query",
            "name": "sortBy",
            "required": false,
            "type": "string"
          },
          {
            "default": "ASC",
            "description": "Direction of the sort",
            "enum": [
              "ASC",
              "DESC"
            ],
            "in": "query",
            "name": "direction",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "items": {
                "$ref": "#/definitions/Sample"
              },
              "type": "array"
            }
          }
        },
        "tags": [
          "Samples"
        ],
        "description": "Get all samples of a patient in a study",
        "operationId": "getAllSamplesOfPatientInStudyUsingGET"
      }
    },
...
    "/api/studies/{studyId}/samples": {
      "get": {
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "Study ID e.g. acc_tcga",
            "in": "path",
            "name": "studyId",
            "required": true,
            "type": "string"
          },
          {
            "default": "SUMMARY",
            "description": "Level of detail of the response",
            "enum": [
              "ID",
              "SUMMARY",
              "DETAILED",
              "META"
            ],
            "in": "query",
            "name": "projection",
            "required": false,
            "type": "string"
          },
          {
            "default": 10000000,
            "description": "Page size of the result list",
            "format": "int32",
            "in": "query",
            "maximum": 10000000,
            "minimum": 1,
            "name": "pageSize",
            "required": false,
            "type": "integer"
          },
          {
            "default": 0,
            "description": "Page number of the result list",
            "format": "int32",
            "in": "query",
            "minimum": 0,
            "name": "pageNumber",
            "required": false,
            "type": "integer"
          },
          {
            "description": "Name of the property that the result list is sorted by",
            "enum": [
              "sampleId",
              "sampleType"
            ],
            "in": "query",
            "name": "sortBy",
            "required": false,
            "type": "string"
          },
          {
            "default": "ASC",
            "description": "Direction of the sort",
            "enum": [
              "ASC",
              "DESC"
            ],
            "in": "query",
            "name": "direction",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "items": {
                "$ref": "#/definitions/Sample"
              },
              "type": "array"
            }
          }
        },
        "tags": [
          "Samples"
        ],
        "description": "Get all samples in a study",
        "operationId": "getAllSamplesInStudyUsingGET"
      }
    },
    "/api/studies/{studyId}/samples/{sampleId}": {
      "get": {
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "Study ID e.g. acc_tcga",
            "in": "path",
            "name": "studyId",
            "required": true,
            "type": "string"
          },
          {
            "description": "Sample ID e.g. TCGA-OR-A5J2-01",
            "in": "path",
            "name": "sampleId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Sample"
            }
          }
        },
        "tags": [
          "Samples"
        ],
        "description": "Get a sample in a study",
        "operationId": "getSampleInStudyUsingGET"
      }
    },
...
  },
  "definitions": {
...
    "Sample": {
      "properties": {
        "copyNumberSegmentPresent": {
          "type": "boolean"
        },
        "patientId": {
          "type": "string"
        },
        "sampleId": {
          "type": "string"
        },
        "sampleType": {
          "enum": [
            "Primary Solid Tumor",
            "Recurrent Solid Tumor",
            "Primary Blood Tumor",
            "Recurrent Blood Tumor",
            "Metastatic",
            "Blood Derived Normal",
            "Solid Tissues Normal"
          ],
          "type": "string"
        },
        "sequenced": {
          "type": "boolean"
        },
        "studyId": {
          "type": "string"
        },
        "uniquePatientKey": {
          "type": "string"
        },
        "uniqueSampleKey": {
          "type": "string"
        }
      },
      "required": [
        "patientId",
        "sampleId",
        "studyId"
      ],
      "type": "object"
    },
...
  },
...
}